vuln_chains/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pyjwt>=2.9.0
pydantic>=2.0.0
httpx>=0.27.0
